Fixed some minor issues with ld/st not setting the proper flags on the right registers

+++ zenes/6502/Cpu6502.h
@@ -21,7 +21,7 @@
 @property uint16_t op2;
 @property uint16_t op3;
 @property uint8_t interruptPeriod;
-@property uint8_t counter;
+@property uint16_t counter;
 @property uint8_t reg_acc;
 @property uint8_t reg_x;
 @property uint8_t reg_y;
